More type refactoring
