+++ docs/src/components/sidebar-preview.astro
@@ -6,7 +6,7 @@ import type {
 } from '../../../packages/starlight/schemas/sidebar';
 import SidebarSublist from '../../../packages/starlight/components/SidebarSublist.astro';
 import type { Badge } from '../../../packages/starlight/schemas/badge';
-import type { SidebarEntry } from '../../../packages/starlight/utils/navigation';
+import type { SidebarEntry } from '../../../packages/starlight/utils/routing/types';
 
 interface Props {
 	config: SidebarConfig;
