+++ packages/starlight/components/SidebarSublist.astro
@@ -1,5 +1,6 @@
 ---
-import { flattenSidebar, type SidebarEntry } from '../utils/navigation';
+import { flattenSidebar } from '../utils/navigation';
+import type { SidebarEntry } from '../utils/routing/types';
 import Icon from '../user-components/Icon.astro';
 import Badge from '../user-components/Badge.astro';
 import SidebarRestorePoint from './SidebarRestorePoint.astro';
